Update <use>'s target reference when adopting to a new document
When the <use> element moves to a new document, it's both possible that
the base URL changes and that the transition is from an inactive to an
active document. In the latter case we need to re-resolve the URL and
potentially attempt to fetch any external resource again.

Do this by calling UpdateTargetReference() in a DidMoveToNewDocument()
override. Also add a check for being in an inactive document to avoid
the unnecessary work in that case, as well as clearing out the resource
reference if the move is from an active to an inactive document.

Fixed: 1070076
Change-Id: Ibfcbe5a590d2115eece0d68ed1c03db49ef5f708